S15.g.1: split runtime adapter into runtimes/openclaw/ package
Mechanical move. Lifts the AzureClaw runtime adapter for OpenClaw out
of cli/src/ into its own top-level package runtimes/openclaw/, sibling
to the future runtimes/openai-agents/ and runtimes/maf/ adapters
(S10.A3 + S10.A4). No behaviour change.

Moves (git mv preserves history):
  cli/src/plugin.ts            → runtimes/openclaw/src/index.ts
  cli/src/core/                → runtimes/openclaw/src/core/
  cli/src/plugin.test.ts       → runtimes/openclaw/src/index.test.ts
  cli/src/redact.test.ts       → runtimes/openclaw/src/redact.test.ts
  cli/src/router-url.test.ts   → runtimes/openclaw/src/router-url.test.ts
  cli/openclaw.plugin.json     → runtimes/openclaw/openclaw.plugin.json

New runtimes/openclaw/{package.json,tsconfig.json,.gitignore} —
@azureclaw/runtime-openclaw, narrowed deps (@agentmesh/sdk +
commander only; operator-CLI-only deps stay in cli/).

Sandbox Dockerfile cli-builder stage repointed to runtimes/openclaw/.
ci/loc-budget.yaml entry repointed. New 'Runtime OpenClaw Build &
Test' CI job at parity with cli-build and mesh-plugin-build.

Verification: cd runtimes/openclaw && tsc / lint / test (100 pass) /
build all clean. cd cli && tsc / lint / test (354 pass / 2 skipped) /
build all clean.
