+++ .github/workflows/ci.yml
@@ -65,6 +65,25 @@ jobs:
       - name: npm audit (CLI dependencies)
         run: npm audit --audit-level=high
 
+  runtime-openclaw-build:
+    name: Runtime OpenClaw Build & Test
+    runs-on: ubuntu-latest
+    defaults:
+      run:
+        working-directory: runtimes/openclaw
+    steps:
+      - uses: actions/checkout@34e114876b0b11c390a56381ad16ebd13914f8d5 # v4
+      - uses: actions/setup-node@49933ea5288caeca8642d1e84afbd3f7d6820020 # v4
+        with:
+          node-version: "22"
+      - run: npm install
+      - run: npm run typecheck
+      - run: npm run lint
+      - run: npm run build
+      - run: npm test
+      - name: npm audit (Runtime OpenClaw dependencies)
+        run: npm audit --audit-level=high
+
   mesh-plugin-build:
     name: Mesh Plugin Build & Test
     runs-on: ubuntu-latest
