+++ docs/security-audits/2026-04-29-phase2-runtime-package-split-g1.md
@@ -1,0 +1,120 @@
+# 2026-04-29 — Phase 2 / S15.g.1 — runtime adapter package split (`runtimes/openclaw/`)
+
+## Scope
+
+Mechanical move. Lifts the AzureClaw runtime adapter for OpenClaw out
+of `cli/src/` (where it was intermingled with the operator CLI) into
+its own top-level package `runtimes/openclaw/`, sibling to the future
+`runtimes/openai-agents/` and `runtimes/maf/` adapters slated for
+S10.A3 + S10.A4. No behaviour change: every moved file is byte-identical
+or has only its own relative-import path edited (`./plugin.js` →
+`./index.js`).
+
+## Existing implementation surveyed
+
+- `cli/src/plugin.ts` (2463 LOC after the S15.f.10 cap was reached):
+  the OpenClaw plugin definition + `register()` body. Only consumers
+  outside its own subtree were two unit-test files
+  (`cli/src/redact.test.ts`, `cli/src/router-url.test.ts`) importing
+  `redactSecrets` and the `routerUrl` family from `./plugin.js`.
+  These tests test runtime behaviour, not CLI behaviour, so they
+  follow the source file out of `cli/`.
+- `cli/src/core/` (8 modules + `agt-tools/` + `commands/` subdirs,
+  S15.f.5 → S15.f.10 decomposition output): all imports inside the
+  subtree are relative — none cross the `cli/src/core/` boundary
+  outward to other parts of `cli/`.
+- `cli/openclaw.plugin.json`: the OpenClaw plugin manifest. Lists
+  `skills/<name>` paths relative to the plugin install location
+  (`/opt/azureclaw-plugin/`). Manifest is moved with the runtime; the
+  Dockerfile still copies skills next to it (S15.g.2 will move skills
+  out of `cli/skills/` to `skills/foundry/`).
+- `sandbox-images/openclaw/Dockerfile` `cli-builder` stage: previously
+  did `COPY cli/{package.json,package-lock.json,tsconfig.json}` +
+  `COPY cli/src/` + `COPY policy-engine/profiles/` and ran
+  `npm ci && npm run build` there. The `policy-engine/profiles/` copy
+  exists only because `cli/package.json`'s `build` script post-copies
+  those into `cli/dist/profiles/` for use by the host-side
+  `azureclaw dev` command (seccomp staging) — the in-sandbox runtime
+  adapter never reads those profile JSONs. So the new
+  `runtime-openclaw` builder stage drops the `policy-engine/profiles/`
+  copy; cli's own build pipeline still produces them for the operator
+  CLI tarball.
+- No Helm chart, no Kustomize overlay, and no controller code
+  references `cli/dist/plugin.js` or `cli/openclaw.plugin.json`
+  directly. The only consumer of the `dist/` plugin layout is the
+  sandbox Dockerfile.
+
+## File moves (`git mv` to preserve history)
+
+| Old path | New path |
+|---|---|
+| `cli/src/plugin.ts` | `runtimes/openclaw/src/index.ts` |
+| `cli/src/core/` | `runtimes/openclaw/src/core/` |
+| `cli/src/plugin.test.ts` | `runtimes/openclaw/src/index.test.ts` |
+| `cli/src/redact.test.ts` | `runtimes/openclaw/src/redact.test.ts` |
+| `cli/src/router-url.test.ts` | `runtimes/openclaw/src/router-url.test.ts` |
+| `cli/openclaw.plugin.json` | `runtimes/openclaw/openclaw.plugin.json` |
+
+## New files
+
+- `runtimes/openclaw/package.json` (`@azureclaw/runtime-openclaw`,
+  `main: dist/index.js`, `openclaw.extensions: ["./dist/index.js"]`).
+  Deps narrowed to the actual surface: `@agentmesh/sdk` + `commander`.
+  Operator-CLI-only deps (`@azure/identity`, `blessed`, `inquirer`,
+  `node-pty`, `execa`, …) stay in `cli/package.json` only.
+- `runtimes/openclaw/tsconfig.json` mirrors `cli/tsconfig.json`.
+- `runtimes/openclaw/package-lock.json` generated by `npm install`
+  (also committed for reproducible Dockerfile builds).
+
+## Edited references
+
+- `sandbox-images/openclaw/Dockerfile` — `cli-builder` stage rewritten
+  to build `runtimes/openclaw/`; `COPY --from=cli-builder
+  /build/runtimes/openclaw/dist/` and `COPY runtimes/openclaw/
+  package.json runtimes/openclaw/package-lock.json runtimes/openclaw/
+  openclaw.plugin.json /opt/azureclaw-plugin/`. `cli/skills/` +
+  `cli/policies/` paths unchanged for S15.g.1 (the next sub-slice
+  moves skills).
+- `ci/loc-budget.yaml` — `cli/src/plugin.ts` budget entry repointed to
+  `runtimes/openclaw/src/index.ts`.
+- `.github/workflows/ci.yml` — new `Runtime OpenClaw Build & Test`
+  job at parity with `cli-build` and `mesh-plugin-build`. Runs
+  `typecheck / lint / build / test / npm audit` for the new package.
+- Two unit-test files had `./plugin.js` import specifiers rewritten
+  to `./index.js`. Dynamic `import("./plugin.js")` calls inside
+  `index.test.ts` (the moved `plugin.test.ts`) likewise rewritten.
+
+## Verification
+
+- `cd runtimes/openclaw && npx tsc --noEmit` — clean.
+- `cd runtimes/openclaw && npm run lint` — 23 warnings, 0 errors.
+- `cd runtimes/openclaw && npm test -- --run` — 100 passed (3 files).
+- `cd runtimes/openclaw && npm run build` — clean.
+- `cd cli && npx tsc --noEmit` — clean.
+- `cd cli && npm run lint` — 16 warnings, 0 errors.
+- `cd cli && npm test -- --run` — 354 passed / 2 skipped (the
+  −100 delta vs the pre-S15.g.1 baseline of 454 are the three test
+  files that followed the source out of `cli/`).
+- `cd cli && npm run build` — clean.
+
+## Risk + rollback
+
+- Risk: medium. `git mv` preserves history. The functional risk
+  surface is the sandbox Dockerfile path rewrite — sandbox image
+  rebuild + Kind smoke is the natural next verification (S16
+  chaos-tier covers the latter). cli + runtime in-process tests both
+  pass at parity.
+- Rollback: revert this PR. Files return to `cli/src/`; Dockerfile
+  reverts; CI workflow loses the new job.
+
+## §14.6 / §15 impact
+
+None directly. Pure architectural hygiene. Improves OSS-readiness
+(clear separation of "operator tooling" vs "agent runtime
+adapters"). Unblocks S10.A3 / S10.A4 to land under `runtimes/`
+without re-rooting prior file moves.
+
+## Sign-offs
+
+- Implementer: Copilot CLI agent (S15.g.1 ship).
+- Reviewer: pending PR review.
