+++ runtimes/openclaw/src/core/foundry-discovery.ts
