+++ runtimes/openclaw/src/core/safe-json.ts
